fix: resetting walkstyle to crouch (#1)
The issue as described in issue #1 is that the walkstyle get's reset back but before it's applied it collects your walkstyle again (which is crouched) and then saves that for next time you exit crouch. The fix is to not be able to save move_ped_crouched as your walkstyle and not overwriting the saved walkstyle if the new one is nil.